+++ playwright.config.js
@@ -4,6 +4,7 @@ const { defineConfig } = require(`@playwright/test`);
 
 module.exports = defineConfig({
 	testDir: `./tests/e2e`,
+	globalSetup: require.resolve(`./tests/e2e/global-setup.js`),
 	timeout: 30000,
 	expect: {
 		timeout: 10000