+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - `_env.url`
   - `_env.key`
 - `testdomain` variable is now deprecated in favor of `_env.url`
+- Add the ability to always remember the selected environment at the project level
+- Added a new settings modal for managing this new project-level setting, along with the app-wide default for all projects.
 
 # v26.2.34056